feat: fetch pet list from navbar
