+++ src/pages/api/pet/list/[id].ts
@@ -11,21 +11,11 @@ export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse<ResponseType>
 ) {
-  const token = req.headers.authorization;
-
-  if (!token) {
-    return res.status(401).json({ message: "請重新登入" });
-  }
-
   const id = req.query.id;
 
   try {
     const response = await fetch(`${apiBase.GET_PET_LIST}/${id}`, {
       method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: token,
-      },
     });
 
     const result = await response.json();
